# src/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RVM API",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/volunteers": {
            "get": {
                "tags": [
                    "Volunteers"
                ],
                "summary": "Return all volunteers",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Volunteers"
                ],
                "summary": "Create volunteer",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "organisation_id",
                        "in": "query",
                        "description": "Volunteer organisation id.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Volunteer name.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "ssn",
                        "in": "query",
                        "description": "Volunteer ssn.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Volunteer email.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Volunteer phone.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "county",
                        "in": "query",
                        "description": "Volunteer county.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Volunteer city.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Volunteer address.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "comments",
                        "in": "query",
                        "description": "Volunteer comments.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "job",
                        "in": "query",
                        "description": "Volunteer job.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "added_by",
                        "in": "query",
                        "description": "Volunteer added by.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/volunteers/{id}": {
            "get": {
                "tags": [
                    "Volunteers"
                ],
                "summary": "Show volunteer info ",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Volunteers"
                ],
                "summary": "Update volunteer",
                "operationId": "update",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Volunteers"
                ],
                "summary": "Delete volunteer",
                "operationId": "delete",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
...
}